Time: 14 ms (5.74%) | Memory: 18.3 MB (5.43%) - LeetSync

+++ 1221-element-appearing-more-than-25-in-sorted-array/element-appearing-more-than-25-in-sorted-array.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int findSpecialInteger(vector<int>& arr) {
-        std::map<int,int> values;
+        std::map<int,short> values;
 
         for(int i = 0 ; i < arr.size();++i){
